Amend openCRX backend

+++ src/java/org/opentdc/rates/RatesService.java
@@ -79,8 +79,8 @@ public class RatesService extends GenericService<ServiceProvider> {
 	public List<RatesModel> list(
 		@DefaultValue(DEFAULT_QUERY_TYPE) @QueryParam("queryType") String queryType,
 		@DefaultValue(DEFAULT_QUERY) @QueryParam("query") String query,
-		@DefaultValue(DEFAULT_POSITION) @QueryParam("position") long position,
-		@DefaultValue(DEFAULT_SIZE) @QueryParam("size") long size
+		@DefaultValue(DEFAULT_POSITION) @QueryParam("position") int position,
+		@DefaultValue(DEFAULT_SIZE) @QueryParam("size") int size
 	) {
 		return sp.list(queryType, query, position, size);
 	}
